Extract pull request components into their own files
